+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- ref(js): Remove Logs opt-in step from JS wizards and rely on the SDK's default `enableLogs` ([#1329](https://github.com/getsentry/sentry-wizard/pull/1329))
+
 ## 7.0.2
 
 ### Fixes